+++ noodles-fasta/src/writer/builder.rs
@@ -5,22 +5,11 @@ use super::Writer;
 const DEFAULT_LINE_BASE_COUNT: usize = 80;
 
 /// A FASTA writer builder.
-pub struct Builder<W> {
-    inner: W,
+pub struct Builder {
     line_base_count: usize,
 }
 
-impl<W> Builder<W>
-where
-    W: Write,
-{
-    pub(super) fn new(inner: W) -> Self {
-        Builder {
-            inner,
-            line_base_count: DEFAULT_LINE_BASE_COUNT,
-        }
-    }
-
+impl Builder {
     /// Sets the number of bases per line.
     ///
     /// By default, this is set to 80.
@@ -29,25 +18,37 @@ where
     ///
     /// ```
     /// use noodles_fasta as fasta;
-    /// let builder = fasta::Writer::builder(Vec::new()).set_line_base_count(100);
+    /// let builder = fasta::writer::Builder::default().set_line_base_count(100);
     /// ```
     pub fn set_line_base_count(mut self, line_base_count: usize) -> Self {
         self.line_base_count = line_base_count;
         self
     }
 
-    /// Builds a FASTA writer.
+    /// Builds a FASTA writer from a writer.
     ///
     /// # Examples
     ///
     /// ```
+    /// # use std::io;
     /// use noodles_fasta as fasta;
-    /// let writer = fasta::Writer::builder(Vec::new()).build();
+    /// let writer = fasta::writer::Builder::default().build_with_writer(io::sink());
     /// ```
-    pub fn build(self) -> Writer<W> {
+    pub fn build_with_writer<W>(self, writer: W) -> Writer<W>
+    where
+        W: Write,
+    {
         Writer {
-            inner: self.inner,
+            inner: writer,
             line_base_count: self.line_base_count,
+        }
+    }
+}
+
+impl Default for Builder {
+    fn default() -> Self {
+        Self {
+            line_base_count: DEFAULT_LINE_BASE_COUNT,
         }
     }
 }